Use 64-bit index for JPEG coefficient blocks to fix large-image lossless reconstruction

+++ lib/jxl/jpeg/dec_jpeg_data_writer.cc
@@ -843,7 +843,8 @@ SerializationStatus JXL_NOINLINE DoEncodeScan(const JPEGData& jpg,
           for (int ix = 0; ix < n_blocks_x; ++ix) {
             int block_y = ss.mcu_y * n_blocks_y + iy;
             int block_x = mcu_x * n_blocks_x + ix;
-            int block_idx = block_y * c.width_in_blocks + block_x;
+            size_t block_idx =
+                static_cast<size_t>(block_y) * c.width_in_blocks + block_x;
             if (ss.block_scan_index == ss.next_reset_point) {
               Flush(coding_state, bw);
               ss.next_reset_point = get_next_reset_point();
